Add docs, test in more schemas
