+++ guides/guides.html
@@ -10,6 +10,7 @@ sections:
   - name: Errors
   - name: Pagination
   - name: Relay
+  - name: Dataloader
   - name: Subscriptions
   - name: GraphQL Pro
   - name: GraphQL Pro - OperationStore
