+++ spec/support/dummy/schema.rb
@@ -528,6 +528,7 @@ module Dummy
     use GraphQL::Execution::Interpreter
     use GraphQL::Analysis::AST
     use GraphQL::Execution::Errors
+    use GraphQL::Dataloader
     lazy_resolve(Proc, :call)
   end
