+++ spec/support/jazz.rb
@@ -948,6 +948,7 @@ module Jazz
     if TESTING_INTERPRETER
       use GraphQL::Execution::Interpreter
       use GraphQL::Analysis::AST
+      use GraphQL::Dataloader
     end
   end
